Removed old rest.py and replaced it with the new one.
